+++ src/reactpy/logging.py
@@ -18,12 +18,7 @@ dictConfig(
                 "stream": sys.stdout,
             }
         },
-        "formatters": {
-            "generic": {
-                "format": "%(asctime)s | %(log_color)s%(levelname)s%(reset)s | %(message)s",
-                "datefmt": r"%Y-%m-%dT%H:%M:%S%z",
-            }
-        },
+        "formatters": {"generic": {"datefmt": r"%Y-%m-%dT%H:%M:%S%z"}},
     }
 )